Factor out DiscoveredParticipant, DiscoveredPublication, DiscoveredSubscription
Problem
-------

Sedp.h is too big.

Solution
--------

Make Sedp.h smaller by factoring out data structures.
